+++ api/src/main/java/org/io/rideout/resource/UserResource.java
@@ -11,21 +11,25 @@ import io.swagger.v3.oas.annotations.security.SecurityRequirement;
 import org.bson.types.ObjectId;
 import org.io.rideout.BeanValidation;
 import org.io.rideout.PasswordManager;
+import org.io.rideout.authentication.AuthenticationFilter;
 import org.io.rideout.authentication.Secured;
 import org.io.rideout.database.UserDao;
-import org.io.rideout.database.VehicleDao;
+import org.io.rideout.exception.UnauthorizedException;
 import org.io.rideout.model.User;
-import org.io.rideout.model.Vehicle;
 
 import javax.ws.rs.*;
+import javax.ws.rs.core.Context;
 import javax.ws.rs.core.MediaType;
-import java.util.ArrayList;
+import javax.ws.rs.core.SecurityContext;
+import java.util.Collections;
+import java.util.List;
+
+import static org.io.rideout.authentication.AuthenticationFilter.AUTHENTICATION_SCHEMA;
 
 @Path("user")
 public class UserResource {
 
     private UserDao userDao = UserDao.getInstance();
-    private VehicleDao vehicleDao = VehicleDao.getInstance();
 
     // GET all users
     @GET
@@ -45,8 +49,16 @@ public class UserResource {
                     name = "JWT"
             )
     )
-    public ArrayList<User> getAllUsers() {
-        return userDao.getAll();
+    public List<User> getAllUsers(@Context SecurityContext securityContext) {
+        if (securityContext.isUserInRole(User.STAFF)) {
+            return userDao.getAll();
+        } else if (securityContext.isUserInRole(User.RIDER)) {
+            User result = userDao.getById(new ObjectId(securityContext.getUserPrincipal().getName()));
+            if (result == null) throw new NotFoundException();
+            return Collections.singletonList(result);
+        }
+
+        throw new UnauthorizedException();
     }
 
     // GET user by ID
@@ -69,7 +81,13 @@ public class UserResource {
                     name = "JWT"
             )
     )
-    public User getUser(@Parameter(description = "User ID", schema = @Schema(type = "string")) @PathParam("id") ObjectId id) {
+    public User getUser(@Parameter(description = "User ID", schema = @Schema(type = "string")) @PathParam("id") ObjectId id,
+                        @Context SecurityContext securityContext) {
+
+        if (securityContext.isUserInRole(User.RIDER) && !id.toHexString().equals(securityContext.getUserPrincipal().getName())) {
+            throw new UnauthorizedException();
+        }
+
         User result = userDao.getById(id);
 
         if (result != null) {
@@ -79,61 +97,9 @@ public class UserResource {
         throw new NotFoundException();
     }
 
-    // GET user vehicles
-    @GET
-    @Secured
-    @Path("{id}/vehicle")
-    @Produces(MediaType.APPLICATION_JSON)
-    @Operation(
-            summary = "Get all Vehicles",
-            tags = {"user", "vehicle"},
-            description = "Returns all Vehicles for given user",
-            responses = {
-                    @ApiResponse(responseCode = "200", description = "List of Vehicles", content = @Content(
-                            array = @ArraySchema(schema = @Schema(implementation = Vehicle.class))
-                    )),
-                    @ApiResponse(responseCode = "401", description = "User unauthorized")
-            },
-            security = @SecurityRequirement(
-                    name = "JWT"
-            )
-    )
-    public ArrayList<Vehicle> getUserVehicles(@Parameter(description = "User ID", schema = @Schema(type = "string")) @PathParam("id") ObjectId id) {
-        ArrayList<Vehicle> result = vehicleDao.getAll(id);
-
-        if (result != null) return result;
-        throw new NotFoundException();
-    }
-
-    // GET user vehicle by ID
-    @GET
-    @Secured
-    @Path("{uid}/vehicle/{vid}")
-    @Produces(MediaType.APPLICATION_JSON)
-    @Operation(
-            summary = "Get Vehicle",
-            tags = {"user", "vehicle"},
-            description = "Returns User with given ID",
-            responses = {
-                    @ApiResponse(responseCode = "200", description = "Vehicle", content = @Content(
-                            schema = @Schema(implementation = Vehicle.class)
-                    )),
-                    @ApiResponse(responseCode = "401", description = "User unauthorized"),
-                    @ApiResponse(responseCode = "404", description = "Vehicle not found")
-            },
-            security = @SecurityRequirement(
-                    name = "JWT"
-            )
-    )
-    public Vehicle getUserVehicle(@Parameter(description = "User ID", schema = @Schema(type = "string")) @PathParam("uid") ObjectId uid,
-                                  @Parameter(description = "Vehicle ID", schema = @Schema(type = "string")) @PathParam("vid") ObjectId vid) {
-        Vehicle result = vehicleDao.getById(uid, vid);
-
-        if (result != null) {
-            return result;
-        }
-
-        throw new NotFoundException();
+    @Path("{uid}/vehicle")
+    public VehicleResource getVehicleResource(@Parameter(description = "User ID", schema = @Schema(type = "string")) @PathParam("uid") ObjectId uid) {
+        return new VehicleResource(uid);
     }
 
     @Path("{uid}/riderinfo")
@@ -164,44 +130,17 @@ public class UserResource {
                     name = "JWT"
             )
     )
-    public User updateUser(User user) {
+    public User updateUser(User user, @Context SecurityContext securityContext) {
+        if (securityContext.isUserInRole(User.RIDER) &&
+                !user.getId().toHexString().equals(securityContext.getUserPrincipal().getName())) {
+            throw new UnauthorizedException();
+        }
+
         BeanValidation.validate(user);
         User result = userDao.update(user);
 
         if (result == null) throw new NotFoundException();
         return result;
-    }
-
-    // UPDATE user vehicle
-    @PUT
-    @Secured
-    @Path("{uid}/vehicle/")
-    @Produces(MediaType.APPLICATION_JSON)
-    @Consumes(MediaType.APPLICATION_JSON)
-    @Operation(
-            summary = "Update Vehicle",
-            tags = {"user", "vehicle"},
-            description = "Updates Vehicle with given ID",
-            responses = {
-                    @ApiResponse(responseCode = "200", description = "Vehicle", content = @Content(
-                            schema = @Schema(implementation = Vehicle.class)
-                    )),
-                    @ApiResponse(responseCode = "401", description = "User unauthorized"),
-                    @ApiResponse(responseCode = "404", description = "Vehicle not found")
-            },
-            requestBody = @RequestBody(description = "Updated Vehicle", content = @Content(
-                    schema = @Schema(implementation = Vehicle.class)
-            )),
-            security = @SecurityRequirement(
-                    name = "JWT"
-            )
-    )
-    public Vehicle updateVehicle(@Parameter(description = "User ID", schema = @Schema(type = "string")) @PathParam("uid") ObjectId uid, Vehicle vehicle) {
-        BeanValidation.validate(vehicle);
-        Vehicle result = vehicleDao.update(vehicle);
-
-        if (result != null) return result;
-        throw new NotFoundException();
     }
 
     // CREATE user
@@ -221,45 +160,31 @@ public class UserResource {
                     schema = @Schema(implementation = User.class)
             ))
     )
-    public User addUser(User user) {
+    public User addUser(User user, @HeaderParam("Authorization") @DefaultValue("") String authHeader) {
         user.setId(new ObjectId());
         user.setPassword(PasswordManager.hashPassword(user.getPassword()));
         BeanValidation.validate(user);
 
-        return userDao.insert(user);
-    }
+        if (user.getRole().equals(User.RIDER)) {
+            return userDao.insert(user);
+        } else if (user.getRole().equals(User.STAFF)) {
+            if (!authHeader.isEmpty()) {
+                String token = authHeader.substring(AUTHENTICATION_SCHEMA.length()).trim();
+                String id = AuthenticationFilter.validateToken(token);
 
-    // CREATE user vehicle
-    @POST
-    @Secured
-    @Path("{uid}/vehicle")
-    @Produces(MediaType.APPLICATION_JSON)
-    @Consumes(MediaType.APPLICATION_JSON)
-    @Operation(
-            summary = "Add Vehicle",
-            tags = {"user", "vehicle"},
-            description = "Adds new Vehicle to a User",
-            responses = {
-                    @ApiResponse(responseCode = "200", description = "Vehicle", content = @Content(
-                            schema = @Schema(implementation = Vehicle.class)
-                    )),
-                    @ApiResponse(responseCode = "401", description = "User unauthorized"),
-                    @ApiResponse(responseCode = "404", description = "Vehicle not found")
-            },
-            requestBody = @RequestBody(description = "New Vehicle", content = @Content(
-                    schema = @Schema(implementation = Vehicle.class)
-            )),
-            security = @SecurityRequirement(
-                    name = "JWT"
-            )
-    )
-    public Vehicle addVehicle(@Parameter(description = "User ID", schema = @Schema(type = "string")) @PathParam("uid") ObjectId uid, Vehicle vehicle) {
-        vehicle.setId(new ObjectId());
-        BeanValidation.validate(vehicle);
-        Vehicle result = vehicleDao.insert(uid, vehicle);
+                if (id != null) {
+                    User auth = userDao.getById(new ObjectId(id));
 
-        if (result != null) return result;
-        throw new InternalServerErrorException();
+                    if (auth != null && auth.getRole().equals(User.STAFF)) {
+                        return userDao.insert(user);
+                    }
+                }
+            }
+
+            throw new UnauthorizedException();
+        }
+
+        throw new BadRequestException();
     }
 
     // DELETE user
@@ -282,38 +207,16 @@ public class UserResource {
                     name = "JWT"
             )
     )
-    public User removeUser(@Parameter(description = "User ID", schema = @Schema(type = "string")) @PathParam("id") ObjectId id) {
+    public User removeUser(@Parameter(description = "User ID", schema = @Schema(type = "string")) @PathParam("id") ObjectId id,
+                           @Context SecurityContext securityContext) {
+
+        if (securityContext.isUserInRole(User.RIDER) && !id.toHexString().equals(securityContext.getUserPrincipal().getName())) {
+            throw new UnauthorizedException();
+        }
+
         User result = userDao.delete(id);
 
         if (result == null) throw new NotFoundException();
         return result;
-    }
-
-    //DELETE user vehicle
-    @DELETE
-    @Secured
-    @Path("{uid}/vehicle/{vid}")
-    @Produces(MediaType.APPLICATION_JSON)
-    @Operation(
-            summary = "Delete Vehicle",
-            tags = {"user", "vehicle"},
-            description = "Deletes a Vehicle",
-            responses = {
-                    @ApiResponse(responseCode = "200", description = "Vehicle", content = @Content(
-                            schema = @Schema(implementation = Vehicle.class)
-                    )),
-                    @ApiResponse(responseCode = "401", description = "User unauthorized"),
-                    @ApiResponse(responseCode = "404", description = "Vehicle not found")
-            },
-            security = @SecurityRequirement(
-                    name = "JWT"
-            )
-    )
-    public Vehicle removeVehicle(@Parameter(description = "User ID", schema = @Schema(type = "string")) @PathParam("uid") ObjectId uid,
-                                 @Parameter(description = "Vehicle ID", schema = @Schema(type = "string")) @PathParam("vid") ObjectId vid) {
-        Vehicle result = vehicleDao.delete(uid, vid);
-
-        if (result != null) return result;
-        throw new NotFoundException();
     }
 }
